--- seminars/s_4/task_8.py
@@ -1,5 +1,0 @@
-# Задание №8
-# ✔ Создайте несколько переменных заканчивающихся и не оканчивающихся на «s».
-# ✔ Напишите функцию, которая при запуске заменяет содержимое переменных
-# оканчивающихся на s (кроме переменной из одной буквы s) на None.
-# ✔ Значения не удаляются, а помещаются в одноимённые переменные без s на конце.
