+++ seminars/sem_4/task_7.py
@@ -3,3 +3,20 @@
 # и списком с доходами и расходами (3-10 чисел) в качестве значения.
 # ✔ Вычислите итоговую прибыль или убыток каждой компании. Если все компании
 # прибыльные, верните истину, а если хотя бы одна убыточная — ложь.
+
+companies = {
+    'audi' : [100, -90, 500, -1000, 800, -1800],
+    'opel' : [-100, 900, -400, 5000, -600],
+    'nissan' : [500, -700, 450, -840, 999]
+}
+
+def comp_result(comp_dict):
+    for k, v in comp_dict.items():
+        res = sum(v)
+        print(res)
+        if res < 0:
+            return False
+
+    return True
+
+print(comp_result(companies))
